pyproject.toml
[tool.poetry]
name = "yupxmusic3"
version = "0.1.0"
description = ""
authors = ["Sophon96 <71684640+Sophon96@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
discord-py = {extras = ["voice"], version = "^2.3.1"}


[tool.poetry.group.dev.dependencies]
yapf = "^0.40.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
